Update STAC up to the 1.0.0 version

+++ modules/core/src/main/scala/com/azavea/stac4s/extensions/eo/package.scala
@@ -6,8 +6,8 @@ import eu.timepit.refined.numeric._
 import eu.timepit.refined.types.numeric.PosInt
 
 package object eo {
-  type Percentage = Int Refined Interval.Closed[W.`0`.T, W.`100`.T]
-  object Percentage extends RefinedTypeOps[Percentage, Int]
+  type Percentage = Double Refined Interval.Closed[W.`0D`.T, W.`100D`.T]
+  object Percentage extends RefinedTypeOps[Percentage, Double]
 
   type BandRange = PosInt
 }
